Added: Default buildComparisonExpression method in StringComparisonPredicate class.

+++ lib/eMapper/Query/Builder/DeleteQueryBuilder.php
@@ -4,11 +4,28 @@ namespace eMapper\Query\Builder;
 use eMapper\Engine\Generic\Driver;
 
 class DeleteQueryBuilder extends QueryBuilder {
+	/**
+	 * Indicates if the current query deletes all entities
+	 * @var boolean
+	 */
+	protected $truncate;
+	
+	public function __construct(ClassProfile $entity, $truncate = false) {
+		parent::__construct($entity);
+		$this->truncate = $truncate;
+	}
+	
 	public function build(Driver $driver, $config = null) {
 		$args = [];
 		
+		//get table name
+		$table = $this->entity->getReferencedTable();
+		
 		//evaluate condition
-		if (isset($this->condition)) {
+		if ($this->truncate) {
+			return [sprintf("DELETE FROM %s", $table), null];
+		}
+		elseif (isset($this->condition)) {
 			$condition = $this->condition->evaluate($this->entity, $args);
 		}
 		elseif (array_key_exists('query.filter', $config) && !empty($config['query.filter'])) {
@@ -21,10 +38,11 @@ class DeleteQueryBuilder extends QueryBuilder {
 			$condition = implode(' AND ', $filters);
 		}
 		
-		//get table name
-		$table = $this->entity->getReferencedTable();
+		if (isset($condition)) {
+			return [sprintf("DELETE FROM %s WHERE %s", $table, $condition), $args];
+		}
 		
-		return [sprintf("DELETE FROM %s WHERE %s", $table, $condition), $args];
+		throw new \RuntimeException("No condition specified for deletion query");
 	}
 }
 ?>
